+++ core/src/main/scala/com/krux/hyperion/HyperionAwsClient.scala
@@ -1,17 +1,37 @@
 package com.krux.hyperion
 
 import scala.annotation.tailrec
-import scala.collection.JavaConversions._
+import scala.collection.JavaConverters._
 
 import com.amazonaws.auth.{ DefaultAWSCredentialsProviderChain, STSAssumeRoleSessionCredentialsProvider }
 import com.amazonaws.regions.{ Region, Regions }
+import com.amazonaws.{ AmazonWebServiceRequest, AmazonServiceException }
 import com.amazonaws.services.datapipeline._
 import com.amazonaws.services.datapipeline.model._
 import com.krux.hyperion.DataPipelineDef._
 import org.slf4j.LoggerFactory
 
 sealed trait HyperionAwsClient {
+
+  val maxRetry = 3
+
   protected lazy val log = LoggerFactory.getLogger(getClass)
+
+  def throttleRetry[A](func: => A, n: Int = 0): A = {
+    if (n < maxRetry)
+      try {
+        func
+      } catch {
+        // use startsWith becase the doc says the error code is called "Throttling" but sometimes
+        // we see "ThrottlingException" instead
+        case e: AmazonServiceException if e.getErrorCode().startsWith("Throttling") && e.getStatusCode == 400 =>
+          log.warn(s"caught exception: ${e.getMessage}\n Retry after 5 seconds...")
+          Thread.sleep(5000)
+          throttleRetry(func, n + 1)
+      }
+    else
+      func
+  }
 
   def getPipelineId: Option[String]
   def createPipeline(force: Boolean): Option[String]
@@ -30,13 +50,13 @@ case class HyperionAwsClientForPipelineId(client: DataPipelineClient, pipelineId
 
   def deletePipeline(): Boolean = {
     log.info(s"Deleting pipeline $pipelineId")
-    client.deletePipeline(new DeletePipelineRequest().withPipelineId(pipelineId))
+    throttleRetry(client.deletePipeline(new DeletePipelineRequest().withPipelineId(pipelineId)))
     true
   }
 
   def activatePipeline(): Boolean = {
     log.info(s"Activating pipeline $pipelineId")
-    client.activatePipeline(new ActivatePipelineRequest().withPipelineId(pipelineId))
+    throttleRetry(client.activatePipeline(new ActivatePipelineRequest().withPipelineId(pipelineId)))
     true
   }
 
@@ -51,9 +71,10 @@ case class HyperionAwsClientForName(client: DataPipelineClient, pipelineName: St
         request: ListPipelinesRequest = new ListPipelinesRequest()
       ): List[String] = {
 
-      val response = client.listPipelines(request)
+      val response = throttleRetry(client.listPipelines(request))
 
       val theseIds: List[String] = response.getPipelineIdList
+        .asScala
         .collect { case idName if idName.getName == pipelineName => idName.getId }
         .toList
 
@@ -113,25 +134,34 @@ case class HyperionAwsClientForPipelineDef(client: DataPipelineClient, pipelineD
         }
 
       case None =>
-        val pipelineId = client.createPipeline(
-          new CreatePipelineRequest()
-            .withUniqueId(pipelineDef.pipelineName)
-            .withName(pipelineDef.pipelineName)
-            .withTags(pipelineDef.tags.map { case (k, v) => new Tag().withKey(k).withValue(v.getOrElse("")) })
-        ).getPipelineId
+        val pipelineId = throttleRetry(
+            client.createPipeline(
+              new CreatePipelineRequest()
+                .withUniqueId(pipelineDef.pipelineName)
+                .withName(pipelineDef.pipelineName)
+                .withTags(
+                  pipelineDef.tags
+                    .toSeq
+                    .map { case (k, v) => new Tag().withKey(k).withValue(v.getOrElse("")) }
+                    .asJava
+                )
+            )
+          ).getPipelineId
 
         log.info(s"Pipeline created: $pipelineId")
         log.info("Uploading pipeline definition")
 
-        val putDefinitionResult = client.putPipelineDefinition(
-          new PutPipelineDefinitionRequest()
-            .withPipelineId(pipelineId)
-            .withPipelineObjects(pipelineObjects)
-            .withParameterObjects(parameterObjects)
+        val putDefinitionResult = throttleRetry(
+          client.putPipelineDefinition(
+            new PutPipelineDefinitionRequest()
+              .withPipelineId(pipelineId)
+              .withPipelineObjects(pipelineObjects.asJava)
+              .withParameterObjects(parameterObjects.asJava)
+          )
         )
 
-        putDefinitionResult.getValidationErrors.flatMap(_.getErrors).foreach(log.error)
-        putDefinitionResult.getValidationWarnings.flatMap(_.getWarnings).foreach(log.warn)
+        putDefinitionResult.getValidationErrors.asScala.flatMap(_.getErrors.asScala).foreach(log.error)
+        putDefinitionResult.getValidationWarnings.asScala.flatMap(_.getWarnings.asScala).foreach(log.warn)
 
         if (putDefinitionResult.getErrored) {
           log.error("Failed to create pipeline")